spark/requirements.txt
asyncio==3.4.3
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
gevent==23.9.0.post1
greenlet==2.0.2
idna==3.6
kafka==1.3.5
psycopg2-binary==2.9.9
py4j==0.10.9.7
pycparser==2.21
pyspark==3.5.0
requests==2.31.0
urllib3==1.26.18
websocket==0.2.1
websocket-client==1.6.2
websockets==11.0.3
zope.event==5.0
zope.interface==6.0
